app/requirements.txt
fastapi
uvicorn
python-dotenv
stripe
pydantic
